mqtt-server/requirements.txt
paho-mqtt==1.5.1
pytest==6.2.1
autopep8==1.5
Flask==1.1.2
bokeh==2.2.3